+++ lib/sequentia/__init__.py
@@ -1,10 +1,11 @@
 from .datasets import *
 from .models import *
-# from .preprocessing import *
+from .preprocessing import *
+from .pipeline import *
 from .utils import *
 
 __name__ = "sequentia"
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 __author__ = "Edwin Onuonga"
 __email__ = "ed@eonu.net"
 __copyright__ = f"2019-2023, {__author__}"